Similarly, relocate Windows virtual drive (subst management)
